# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Query Assistant API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@yourcompany.com"
        },
        "version": "1.0"
    },
    "host": "dev.smartbet.live",
    "basePath": "/query-assistant/api/v1",
    "paths": {
        "/banner/generate": {
            "post": {
                "description": "Accepts a banner generation request and returns a generation ID immediately.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "summary": "Start async banner image generation",
                "parameters": [
                    {
                        "description": "Generation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BannerGenerationRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.BannerGenerationAccepted"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/banner/generate/{id}": {
            "get": {
                "description": "Returns status, variants (URLs / b64) and failure info for a generation job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "banner"
                ],
                "summary": "Get banner generation status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Generation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BannerGenerationResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BannerGenerationAccepted": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "generation_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "in_progress"
                }
            }
        },
        "models.BannerGenerationRequest": {
            "type": "object",
            "required": [
                "inputs",
                "language",
                "output"
            ],
            "properties": {
                "inputs": {
                    "$ref": "#/definitions/models.BannerInputs"
                },
                "language": {
                    "description": "e.g. \"en\"",
                    "type": "string"
                },
                "output": {
                    "$ref": "#/definitions/models.BannerOutputParams"
                }
            }
        },
        "models.BannerGenerationResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "failure_reason": {
                    "type": "string",
                    "example": "dall-e-3 generation failed: rate limit"
                },
                "generation_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:05Z"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BannerVariant"
                    }
                }
            }
        },
        "models.BannerInputs": {
            "type": "object",
            "properties": {
                "creative_description": {
                    "type": "string",
                    "example": "Gold coins and dramatic lighting"
                },
                "cta_text": {
                    "type": "string",
                    "example": "Claim Now"
                },
                "headline": {
                    "type": "string",
                    "example": "Get 100% Bonus on First Deposit"
                },
                "secondary_text": {
                    "type": "string",
                    "example": "Up to $500 matched. T\u0026Cs apply."
                },
                "visual_style": {
                    "type": "string",
                    "example": "dark_luxury"
                }
            }
        },
        "models.BannerOutputParams": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string",
                    "example": "url"
                },
                "height": {
                    "type": "integer",
                    "example": 1024
                },
                "max_size_kb": {
                    "type": "integer",
                    "example": 2048
                },
                "width": {
                    "type": "integer",
                    "example": 1792
                }
            }
        },
        "models.BannerVariant": {
            "type": "object",
            "properties": {
                "b64_json": {
                    "type": "string"
                },
                "index": {
                    "type": "integer",
                    "example": 0
                },
                "revised_prompt": {
                    "type": "string",
                    "example": "A high-quality promotional banner..."
                },
                "url": {
                    "type": "string",
                    "example": "https://oaidalleapiprodscus.blob.core.windows.net/..."
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Query validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "Query contains forbidden operation: DROP"
                },
                "query_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
    },
...
}